backend/requirements.txt
fastapi
uvicorn
scikit-learn
numpy
torch
flask
python-multipart
requests
pandas
joblib
yara-python
